Change the Student payment modification when edit a student

+++ classtune/app/controllers/student_controller.rb
@@ -1865,55 +1865,55 @@ class StudentController < ApplicationController
         unless params[:student][:image_file].size.to_f > 280000
           if @student.update_attributes(params[:student])
             
-            abort(@previous_category_id.to_s + "  " + @student.student_category_id.to_s)
-            if @previous_category_id != @student.student_category_id
-              @finance_fees = FinanceFee.find_all_by_student_id(@student.id)
-              @student_fees = @finance_fees.map{|s| s.fee_collection_id}
+            if MultiSchool.current_school.id == 352
+              if @previous_category_id != @student.student_category_id
+                @finance_fees = FinanceFee.find_all_by_student_id(@student.id)
+                @student_fees = @finance_fees.map{|s| s.fee_collection_id}
 
-              @payed_fees=FinanceFee.find(:all,:joins=>"INNER JOIN fee_transactions on fee_transactions.finance_fee_id=finance_fees.id INNER JOIN finance_fee_collections on finance_fee_collections.id=finance_fees.fee_collection_id",:conditions=>"finance_fees.student_id=#{@student.id}",:select=>"finance_fees.fee_collection_id").map{|s| s.fee_collection_id}
-              @payed_fees ||= []
+                @payed_fees=FinanceFee.find(:all,:joins=>"INNER JOIN fee_transactions on fee_transactions.finance_fee_id=finance_fees.id INNER JOIN finance_fee_collections on finance_fee_collections.id=finance_fees.fee_collection_id",:conditions=>"finance_fees.student_id=#{@student.id}",:select=>"finance_fees.fee_collection_id").map{|s| s.fee_collection_id}
+                @payed_fees ||= []
 
-              @fee_collection_dates =FinanceFeeParticular.find(:all,:joins=>"INNER JOIN collection_particulars on collection_particulars.finance_fee_particular_id=finance_fee_particulars.id INNER JOIN finance_fee_collections on finance_fee_collections.id=collection_particulars.finance_fee_collection_id",:conditions=>"finance_fee_particulars.batch_id='#{@student.batch_id}' and finance_fee_particulars.receiver_type='Batch'",:select=>"finance_fee_collections.*").uniq
-              @fee_collection_dates.each do |date|
-                d = FinanceFeeCollection.find(date.id)
-                if @student_fees.include?(d.id)
-                  fee = FinanceFee.find_by_student_id_and_fee_collection_id_batch_id_and_is_paid(@student.id, d.id, @student.batch_id, false)
-                  unless fee.blank?
-                    s = Student.find(@student.id)
-                    FinanceFee.update_student_fee(d,s, fee)
-                  else
-                    s = Student.find(@student.id)
-                    FinanceFee.new_student_fee(d,s)
+                @fee_collection_dates =FinanceFeeParticular.find(:all,:joins=>"INNER JOIN collection_particulars on collection_particulars.finance_fee_particular_id=finance_fee_particulars.id INNER JOIN finance_fee_collections on finance_fee_collections.id=collection_particulars.finance_fee_collection_id",:conditions=>"finance_fee_particulars.batch_id='#{@student.batch_id}' and finance_fee_particulars.receiver_type='Batch'",:select=>"finance_fee_collections.*").uniq
+                @fee_collection_dates.each do |date|
+                  d = FinanceFeeCollection.find(date.id)
+                  if @student_fees.include?(d.id)
+                    fee = FinanceFee.find_by_student_id_and_fee_collection_id_batch_id_and_is_paid(@student.id, d.id, @student.batch_id, false)
+                    unless fee.blank?
+                      s = Student.find(@student.id)
+                      FinanceFee.update_student_fee(d,s, fee)
+                    else
+                      s = Student.find(@student.id)
+                      FinanceFee.new_student_fee(d,s)
+                    end
                   end
                 end
               end
-            end
-            
-            if @previous_batch_id != @student.batch_id
-              @finance_fees = FinanceFee.find_all_by_student_id(@student.id)
-              @student_fees = @finance_fees.map{|s| s.fee_collection_id}
 
-              @payed_fees=FinanceFee.find(:all,:joins=>"INNER JOIN fee_transactions on fee_transactions.finance_fee_id=finance_fees.id INNER JOIN finance_fee_collections on finance_fee_collections.id=finance_fees.fee_collection_id",:conditions=>"finance_fees.student_id=#{@student.id}",:select=>"finance_fees.fee_collection_id").map{|s| s.fee_collection_id}
-              @payed_fees ||= []
+              if @previous_batch_id != @student.batch_id
+                @finance_fees = FinanceFee.find_all_by_student_id(@student.id)
+                @student_fees = @finance_fees.map{|s| s.fee_collection_id}
 
-              @fee_collection_dates =FinanceFeeParticular.find(:all,:joins=>"INNER JOIN collection_particulars on collection_particulars.finance_fee_particular_id=finance_fee_particulars.id INNER JOIN finance_fee_collections on finance_fee_collections.id=collection_particulars.finance_fee_collection_id",:conditions=>"finance_fee_particulars.batch_id='#{@student.batch_id}' and finance_fee_particulars.receiver_type='Batch'",:select=>"finance_fee_collections.*").uniq
-              @fee_collection_dates.each do |date|
-                d = FinanceFeeCollection.find(date.id)
-                if @student_fees.include?(d.id)
-                  fee = FinanceFee.find_by_student_id_and_fee_collection_id_batch_id_and_is_paid(@student.id, d.id, @previous_batch_id, false)
-                  fee.destroy if fee.finance_transactions.empty?
-                  fee = FinanceFee.find_by_student_id_and_fee_collection_id_batch_id_and_is_paid(@student.id, d.id, @student.batch_id, false)
-                  unless fee.blank?
-                    s = Student.find(@student.id)
-                    FinanceFee.update_student_fee(d,s, fee)
-                  else
-                    s = Student.find(@student.id)
-                    FinanceFee.new_student_fee(d,s)
+                @payed_fees=FinanceFee.find(:all,:joins=>"INNER JOIN fee_transactions on fee_transactions.finance_fee_id=finance_fees.id INNER JOIN finance_fee_collections on finance_fee_collections.id=finance_fees.fee_collection_id",:conditions=>"finance_fees.student_id=#{@student.id}",:select=>"finance_fees.fee_collection_id").map{|s| s.fee_collection_id}
+                @payed_fees ||= []
+
+                @fee_collection_dates =FinanceFeeParticular.find(:all,:joins=>"INNER JOIN collection_particulars on collection_particulars.finance_fee_particular_id=finance_fee_particulars.id INNER JOIN finance_fee_collections on finance_fee_collections.id=collection_particulars.finance_fee_collection_id",:conditions=>"finance_fee_particulars.batch_id='#{@student.batch_id}' and finance_fee_particulars.receiver_type='Batch'",:select=>"finance_fee_collections.*").uniq
+                @fee_collection_dates.each do |date|
+                  d = FinanceFeeCollection.find(date.id)
+                  if @student_fees.include?(d.id)
+                    fee = FinanceFee.find_by_student_id_and_fee_collection_id_batch_id_and_is_paid(@student.id, d.id, @previous_batch_id, false)
+                    fee.destroy if fee.finance_transactions.empty?
+                    fee = FinanceFee.find_by_student_id_and_fee_collection_id_batch_id_and_is_paid(@student.id, d.id, @student.batch_id, false)
+                    unless fee.blank?
+                      s = Student.find(@student.id)
+                      FinanceFee.update_student_fee(d,s, fee)
+                    else
+                      s = Student.find(@student.id)
+                      FinanceFee.new_student_fee(d,s)
+                    end
                   end
                 end
               end
-            end
-            
+            end 
             username = MultiSchool.current_school.code.to_s+"-"+@student.admission_no        
             champs21_api_config = YAML.load_file("#{RAILS_ROOT.to_s}/config/champs21.yml")['champs21']
             api_endpoint = champs21_api_config['api_url']
@@ -1937,44 +1937,56 @@ class StudentController < ApplicationController
         end
       else
         if @student.update_attributes(params[:student])
-          
-          if @previous_category_id != @student.student_category_id
-            @finance_fees = FinanceFee.find_all_by_student_id(@student.id)
-            @student_fees = @finance_fees.map{|s| s.fee_collection_id}
-            
-            @fee_collection_dates =FinanceFeeParticular.find(:all,:joins=>"INNER JOIN collection_particulars on collection_particulars.finance_fee_particular_id=finance_fee_particulars.id INNER JOIN finance_fee_collections on finance_fee_collections.id=collection_particulars.finance_fee_collection_id",:conditions=>"finance_fee_particulars.batch_id='#{@student.batch_id}' and finance_fee_particulars.receiver_type='Batch'",:select=>"finance_fee_collections.*").uniq
-            @fee_collection_dates.each do |date|
-              d = FinanceFeeCollection.find(date.id)
-              if @student_fees.include?(d.id)
-                fee = FinanceFee.find_by_student_id_and_fee_collection_id_and_batch_id_and_is_paid(@student.id, d.id, @student.batch_id, false)
-                unless fee.blank?
-                    s = Student.find(@student.id)
-                    FinanceFee.update_student_fee(d,s, fee)
-                else
-                  s = Student.find(@student.id)
-                  FinanceFee.new_student_fee(d,s)
+          found_paid_fees = false
+          paid_fees_type = ""
+          @fee_ids = []
+          fee_ind = 0
+          if MultiSchool.current_school.id == 352
+            if @previous_category_id != @student.student_category_id
+              @finance_fees = FinanceFee.find_all_by_student_id(@student.id)
+              @student_fees = @finance_fees.map{|s| s.fee_collection_id}
+
+              @fee_collection_dates =FinanceFeeParticular.find(:all,:joins=>"INNER JOIN collection_particulars on collection_particulars.finance_fee_particular_id=finance_fee_particulars.id INNER JOIN finance_fee_collections on finance_fee_collections.id=collection_particulars.finance_fee_collection_id",:conditions=>"finance_fee_particulars.batch_id='#{@student.batch_id}' and finance_fee_particulars.receiver_type='Batch'",:select=>"finance_fee_collections.*").uniq
+              @fee_collection_dates.each do |date|
+                d = FinanceFeeCollection.find(date.id)
+                if @student_fees.include?(d.id)
+                  fee = FinanceFee.find_by_student_id_and_fee_collection_id_and_batch_id_and_is_paid(@student.id, d.id, @student.batch_id, false)
+                  unless fee.blank?
+                      s = Student.find(@student.id)
+                      FinanceFee.update_student_fee(d,s, fee)
+                  else
+                    fee = FinanceFee.find_by_student_id_and_fee_collection_id_and_batch_id_and_is_paid(@student.id, d.id, @student.batch_id, true)
+                    unless fee.blank?
+                      found_paid_fees = true
+                      paid_fees_type = "category"
+                      @fee_ids[fee_ind] = fee.id
+                    else
+                      s = Student.find(@student.id)
+                      FinanceFee.new_student_fee(d,s)
+                    end
+                  end
                 end
               end
             end
-          end
 
-          if @previous_batch_id != @student.batch_id
-            @finance_fees = FinanceFee.find_all_by_student_id(@student.id)
-            @student_fees = @finance_fees.map{|s| s.fee_collection_id}
+            if @previous_batch_id != @student.batch_id
+              @finance_fees = FinanceFee.find_all_by_student_id(@student.id)
+              @student_fees = @finance_fees.map{|s| s.fee_collection_id}
 
-            @fee_collection_dates =FinanceFeeParticular.find(:all,:joins=>"INNER JOIN collection_particulars on collection_particulars.finance_fee_particular_id=finance_fee_particulars.id INNER JOIN finance_fee_collections on finance_fee_collections.id=collection_particulars.finance_fee_collection_id",:conditions=>"finance_fee_particulars.batch_id='#{@student.batch_id}' and finance_fee_particulars.receiver_type='Batch'",:select=>"finance_fee_collections.*").uniq
-            @fee_collection_dates.each do |date|
-              d = FinanceFeeCollection.find(date.id)
-              if @student_fees.include?(d.id)
-                fee = FinanceFee.find_by_student_id_and_fee_collection_id_and_batch_id_and_is_paid(@student.id, d.id, @previous_batch_id, false)
-                fee.destroy if fee.finance_transactions.empty?
-                fee = FinanceFee.find_by_student_id_and_fee_collection_id_and_batch_id_and_is_paid(@student.id, d.id, @student.batch_id, false)
-                unless fee.blank?
-                  s = Student.find(@student.id)
-                  FinanceFee.update_student_fee(d,s, fee)
-                else
-                  s = Student.find(@student.id)
-                  FinanceFee.new_student_fee(d,s)
+              @fee_collection_dates =FinanceFeeParticular.find(:all,:joins=>"INNER JOIN collection_particulars on collection_particulars.finance_fee_particular_id=finance_fee_particulars.id INNER JOIN finance_fee_collections on finance_fee_collections.id=collection_particulars.finance_fee_collection_id",:conditions=>"finance_fee_particulars.batch_id='#{@student.batch_id}' and finance_fee_particulars.receiver_type='Batch'",:select=>"finance_fee_collections.*").uniq
+              @fee_collection_dates.each do |date|
+                d = FinanceFeeCollection.find(date.id)
+                if @student_fees.include?(d.id)
+                  fee = FinanceFee.find_by_student_id_and_fee_collection_id_and_batch_id_and_is_paid(@student.id, d.id, @previous_batch_id, false)
+                  fee.destroy if fee.finance_transactions.empty?
+                  fee = FinanceFee.find_by_student_id_and_fee_collection_id_and_batch_id_and_is_paid(@student.id, d.id, @student.batch_id, false)
+                  unless fee.blank?
+                    s = Student.find(@student.id)
+                    FinanceFee.update_student_fee(d,s, fee)
+                  else
+                    s = Student.find(@student.id)
+                    FinanceFee.new_student_fee(d,s)
+                  end
                 end
               end
             end
@@ -1991,7 +2003,11 @@ class StudentController < ApplicationController
           @auth_response = JSON::parse(auth_res.body)
           
           flash[:notice] = "#{t('flash3')}"
-          redirect_to :controller => "student", :action => @action_name_main, :id => @student.id
+          if found_paid_fees
+            redirect_to :controller => "student", :action => "adjust_paid_fees", :id => @student.id, :paid_fees => @fee_ids, :paid_fees_type => paid_fees_type
+          else
+            redirect_to :controller => "student", :action => @action_name_main, :id => @student.id
+          end
         else
           @classes = Course.find(:all, :conditions => ["course_name LIKE ?",params[:student][:class_name]])
           @selected_section = params[:student][:section]
@@ -2023,6 +2039,14 @@ class StudentController < ApplicationController
       batches = Batch.find(:all, :conditions => ["name = ?", batch_name]).map{|b| b.course_id}
       @courses = Course.find(:all, :conditions => ["id IN (?)", batches], :group => "course_name", :select => "course_name", :order => "cast(replace(course_name, 'Class ', '') as SIGNED INTEGER) asc")
       
+    end
+  end
+  
+  def adjust_paid_fees
+    @student = Student.find(params[:id])
+    unless params[:paid_fees_type].nil?
+      @paid_fees_type = params[:paid_fees_type]
+      @paid_fees = params[:paid_fees]
     end
   end
